chore: split util/pod.go by function
